use /gridpoints endpoint
