+++ web/modules/weather_data/src/Service/UnitConversionTrait.php
@@ -13,6 +13,34 @@ trait UnitConversionTrait
         return $mm * 0.03937008;
     }
 
+    public function getSpeedScalar(\stdClass $speed, bool $inMph = true)
+    {
+        $rawValue = $speed->value;
+
+        if ($rawValue == null) {
+            return null;
+        }
+
+        $isKmh = $speed->unitCode == "wmoUnit:km_h-1";
+
+        $out = null;
+        if ($isKmh) {
+            if ($inMph) {
+                $out = $rawValue * 0.6213712;
+            } else {
+                $out = $rawValue;
+            }
+        } else {
+            if ($inMph) {
+                $out = $rawValue;
+            } else {
+                $out = $rawValue / 0.6213712;
+            }
+        }
+
+        return (int) round($out);
+    }
+
     /**
      * Get a temperature scalar from a wmoUnit temperature object.
      */
