+++ web/modules/weather_data/src/Service/WeatherDataService.php
@@ -15,6 +15,7 @@ use Symfony\Component\HttpFoundation\RequestStack;
  */
 class WeatherDataService
 {
+    use HourlyForecastTrait;
     use LoggerChannelTrait;
     use UnitConversionTrait;
     use WeatherAlertTrait;
@@ -877,174 +878,6 @@ class WeatherDataService
         }
 
         return $periods;
-    }
-
-    /**
-     * Get the hourly forecast for a location.
-     *
-     * Note that the $now object should *NOT* be set. It's a dependency injection
-     * hack so we can mock the current date/time.
-     *
-     * @return array
-     *   The hourly forecast as an associative array.
-     */
-    public function getHourlyForecastFromGrid(
-        $wfo,
-        $gridX,
-        $gridY,
-        $now = false,
-        $self = false,
-    ) {
-        if (!$self) {
-            $self = $this;
-        }
-
-        $wfo = strtoupper($wfo);
-        if (!($now instanceof \DateTimeImmutable)) {
-            $now = new \DateTimeImmutable();
-        }
-
-        date_default_timezone_set("America/New_York");
-
-        $place = $self->getPlaceFromGrid($wfo, $gridX, $gridY);
-        $timezone = $place->timezone;
-
-        $forecast = $this->getFromWeatherAPI("/gridpoints/$wfo/$gridX,$gridY")
-            ->properties;
-
-        function getForecastPropertiesAsHourlyPeriods($all, $propertyName)
-        {
-            $periods = [];
-            $property = $all->$propertyName;
-
-            foreach ($property->values as $value) {
-                $time = explode("/", $value->validTime);
-                $date = \DateTimeImmutable::createFromFormat(
-                    \DateTimeInterface::ISO8601_EXPANDED,
-                    $time[0],
-                );
-                $interval = new \DateInterval($time[1]);
-
-                $period = new \DatePeriod(
-                    $date,
-                    \DateInterval::createFromDateString("1 hour"),
-                    $date->add($interval),
-                );
-
-                foreach ($period as $hour) {
-                    $periods[] = [
-                        $propertyName => $value->value,
-
-                        // Set to the start of the hour.
-                        "timestamp" => $hour->setTime(
-                            (int) $hour->format("H"),
-                            0,
-                            0,
-                            0,
-                        ),
-                    ];
-                }
-            }
-
-            return $periods;
-        }
-
-        $getTimestamp = function ($period) {
-            return $period["timestamp"];
-        };
-
-        $units = ["temperature" => $forecast->temperature->uom];
-
-        $properties = ["temperature", "probabilityOfPrecipitation"];
-
-        $units = array_map(function ($property) use ($forecast) {
-            return [$property => $forecast->$property->uom];
-        }, $properties);
-        $units = array_merge(...$units);
-
-        $properties = array_map(function ($property) use ($forecast) {
-            return getForecastPropertiesAsHourlyPeriods($forecast, $property);
-        }, $properties);
-
-        $timestamps = array_map(function ($property) use ($getTimestamp) {
-            return array_map($getTimestamp, $property);
-        }, $properties);
-
-        $periods = array_unique(array_merge(...$timestamps), \SORT_REGULAR);
-
-        // Toss out any time periods in the past.
-        $periods = array_filter($periods, function ($period) use (&$now) {
-            $diff = $now->diff($period, false);
-
-            return $diff->invert != 1;
-        });
-
-        $forecast = array_map(function ($period) use ($properties) {
-            $matches = array_map(function ($property) use ($period) {
-                $index = array_search(
-                    $period,
-                    array_column($property, "timestamp"),
-                );
-                if ($index !== false) {
-                    return $property[$index];
-                }
-                return false;
-            }, $properties);
-
-            $matches = array_filter($matches, function ($match) {
-                return $match !== false;
-            });
-
-            if (count($matches) == count($properties)) {
-                return array_merge(...$matches);
-            }
-            return false;
-        }, $periods);
-
-        $forecast = array_filter($forecast, function ($period) {
-            return $period !== false;
-        });
-
-        // // Now map all those forecast periods into the structure we want.
-        $forecast = array_map(function ($period) use (&$timezone) {
-            // This closure needs access to the $timezone variable about. The easiest
-            // way I found to do it was using it by reference.
-            // From the start period of the time, parse it as an ISO8601 string and
-            // then format it into just the "Hour AM/PM" format (e.g., "8 PM")
-            $timestamp = $period["timestamp"]->setTimeZone(
-                new \DateTimeZone($timezone),
-            );
-            $timeString = $timestamp->format("g A");
-
-            // $obsKey = $this->getApiObservationKey($period);
-
-            // "relativeHumidity" => $period->relativeHumidity->value,
-            // "windSpeed" => $period->windSpeed,
-            // "windDirection" => $period->windDirection,
-            // "dewpoint" => $this->getTemperatureScalar($period->dewpoint),
-
-            return [
-                // "conditions" => $this->t->translate(
-                //     ucfirst(strtolower($period->shortForecast)),
-                // ),
-                // "icon" => $this->legacyMapping->$obsKey->icon,
-                // "iconBasename" => $this->getIconFileBasename($obsKey),
-                "probabilityOfPrecipitation" =>
-                    $period["probabilityOfPrecipitation"],
-                "time" => $timeString,
-                "timestamp" => $period["timestamp"]->format("c"),
-                "temperature" => $this->getTemperatureScalar(
-                    (object) [
-                        "uom" => $units["temperature"],
-                        "value" => $period["temperature"],
-                    ],
-                ),
-            ];
-        }, $forecast);
-
-        // Reindex the array. array_filter maintains indices, so it can result in
-        // holes in the array. Bizarre behavior choice, but okay...
-        return array_values($forecast);
     }
 
     /**
